Expose remix request object to getLoadContext
This patch updates `server.ts` to add a third argument to `getLoadContext` which is the transformed Remix Request object. This is helpful for us at my company as we're switching from Cloudflare Workers to a NodeJS based Remix architecture and much of our existing `getLoadContext` is based off of a `Request` object being available.

I also think this will help others write more portable code as well.